Clean up and simplify specs
Realized that now that we’re using req-un and opt-un, we won’t need to
qualify the keys in the diagram maps, and therefore we don’t need the
“convenience” specs that were really just pointers to more generic
specs.

I thought I was going to need those “convenience” specs because I had
forgotten about req-un and opt-un, so I had been thinking that I’d need
to process each diagram map as parsed from the YAML, with unqualified
keys, in order to qualify the keys. I had been thinking that it’d be
really hard to know which namespace to apply to each key, unless they
were extremely regular — i.e. if every map used the same namespace for
all its keys.

Anyway, turns out, that was all wrong; tldr Rich Hickey is a little
smarter than me.