[tool: automated rendering] DRY testing environments
I noticed a lot of duplication between the CircleCI test job and the
local Dockerfile for running the tests. So I moved some of the duplicated
commands and their comments to scripts in bin. This way it’ll be less work
and less cognitive load to keep local and CI builds in sync.