[tool] Tweak test of automated renderer
The test has been failing for me this morning, and I’ve observed that
the live version of Structurizr Express (SE) _seems_ to have been
changed; it now seems to no longer be including the logo nor the
watermark in the rendered image files. Which is... fine, but a little
surprising. If SE is going to be changing so frequently, then my
thinking is that we need to relax the threshold of difference that's
allowable in the automated test when it compares the “expected” image to
the “actual” image.'