[tool: automated rendering] Fix render test with new approach
The problem *was* that Structurizr Express (SE) puts its watermark in a
random location at a random angle, so I can’t expect its output to be
exactly identical to any kind of pre-saved “expected” value.

The prior attempt to blow out the contrast just wasn’t sufficiently
reliable, and was also problematic in some other ways.

This new approach computes the actual difference between the images, as
in the pixel/color values, and uses that to determine whether the actual
rendered image is sufficiently similar to the expected value. This seems
to work quite well! I tested various modifications to the “expected”
file and any change I made that was notable to my eye was caught by the
test. Pretty happy with this!